test(codenexus): harden generated path exclusions
